requirements.txt
streamlit
langchain
langchain-together
langchain-community
ddgs
python-dotenv